+++ src/actions/CRUD/purchLine/editPurchline.ts
@@ -11,7 +11,7 @@ export async function editPurchline(
             console.log('editPurchline', formData)
             const purchase_id: number = parseInt(formData.get('purchase_id') as string);
             const item_id: number = parseInt(formData.get('item_id') as string);
-            const quantity: number = parseInt(formData.get('quantity') as string);
+            const quantity: number = parseFloat(formData.get('quantity') as string);
             const unit_id: number = parseInt(formData.get('unit_id') as string);
             const id: number = parseInt(formData.get('purch_line_id') as string);
